Refactor HTTP routing to integrate Hono and Convex Auth: replace the previous HTTP actions with a Hono-based router for handling routes, including CORS middleware for AI endpoints. Remove openai_http_actions.ts and update API definitions accordingly to streamline AI functionality. Enhance code organization and maintainability.
